Allow continuation on archive errors. More debug.
This commit allows the user to continue processing data when an archive
is unable to be extracted. This is useful in cases where a single
archive may be corrupt in a large run.

This commit also introduces additional debug messages when STACS is run
with the `--debug` flag, in order to provide more visibility during
development and troubleshooting.

This change has been added to resolve Github issue #12.