Gracefully handle binary data in "text" files
In the case that a file which is believed to be text actually contains
binary data, we'll catch the encoding errors and return the location of
the finding instead of the line number.

This allows corrupt, malformed, and strange file formats which may
present as text but contain binary data to be handled gracefully.

This commit resolves GH-3.